back/requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.1
Pillow==9.2.0
PyJWT==1.7.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.7
